+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.40.0
 
 * Remove support for Apache Kafka 3.5.0 and 3.5.1
+* The `UseKRaft` feature gate moves to beta stage and is enabled by default.
+  If needed, `UseKRaft` can be disabled in the feature gates configuration in the Cluster Operator.
 * Fix NullPointerException from missing listenerConfig when using custom auth
 * Added support for Kafka Exporter `offset.show-all` parameter